+++ src/qconnect/qcmgr.hxx
@@ -107,7 +107,7 @@ private:
     void onTracksInserted(const std::vector<QueueTrack>& tracks,
                            uint32_t insert_after_item_id);
     void onTracksAdded(const std::vector<QueueTrack>& tracks);
-    void onTracksRemoved(const std::vector<uint32_t>& queue_item_ids);
+    void onTracksRemoved(const std::vector<uint64_t>& queue_item_ids);
     void onWsConnected();
     void onWsDisconnected();
 
@@ -119,12 +119,12 @@ private:
     // out_item_ids is filled with the queue_item_id for each returned URL.
     std::vector<std::string> resolveStreamUrls(
         const std::vector<QueueTrack>& tracks,
-        std::vector<uint32_t>& out_item_ids);
+        std::vector<uint64_t>& out_item_ids);
 
     // Look up Qobuz queue_item_id from MPD queue position.
-    uint32_t queueItemIdAt(int mpd_pos) const;
+    uint64_t queueItemIdAt(int mpd_pos) const;
     // Look up MPD queue position from Qobuz queue_item_id. Returns -1 if not found.
-    int mpdPosForQueueItem(uint32_t queue_item_id) const;
+    int mpdPosForQueueItem(uint64_t queue_item_id) const;
 
     // IPC with upmpdcli
     bool startIpcServer();
@@ -147,7 +147,7 @@ private:
 
     // Maps MPD queue position -> Qobuz queue_item_id (parallel to MPD queue)
     mutable std::mutex        m_qmap_mutex;
-    std::vector<uint32_t>     m_queue_item_ids;
+    std::vector<uint64_t>     m_queue_item_ids;
 
     // IPC
     int          m_ipc_sock{-1};
